% Generated by roxygen2 (4.0.0): do not edit by hand
\name{createFigure1}
\alias{createFigure1}
\title{Figure 1}
\usage{
createFigure1()
}
\description{
This function creates Figure 1 in Gibb and Strimmer 2012.
}
\examples{
\dontrun{
library("MALDIquantExamples")
pdfWidth <- 7
pdfHeight <- pdfWidth*0.35

pdf(file="figure1.pdf", height=pdfHeight, width=pdfWidth)
createFigure1()
dev.off()
}
}
\references{
S. Gibb and K. Strimmer. 2012. MALDIquant: a versatile R package for the
analysis of mass spectrometry data. Bioinformatics 28: 2270-2271
}
\seealso{
\code{\link[MALDIquantExamples]{createFigure1Color}}
}

