% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPathNyakas2013.R
\name{getPathNyakas2013}
\alias{getPathNyakas2013}
\title{This dataset contains 2222 MALDI-TOF mass spectra of a mouse kidney.}
\usage{
getPathNyakas2013
}
\value{
Returns the local file path for the corresponding tar-archive.
}
\description{
This MALDI Imaging dataset contains 2222 MALDI-TOF mass spectra of a mouse
kidney. It ranges from (x=29, y=68) to (x=101, y=92).
}
\examples{
library("MALDIquantExamples")
getPathNyakas2013()
}
\references{
This dataset was kindly provided by
Dr. Adrien Nyakas (\email{adrien.nyakas@dcb.unibe.ch}).

See also: \url{http://dx.doi.org/10.6084/m9.figshare.735961}.
}
\keyword{datasets}

