% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPathSpecies}
\alias{getPathSpecies}
\title{This dataset contains 96 MALDI-TOF mass spectra of different bacteria
species.}
\usage{
getPathSpecies
}
\value{
Returns the local file path for the corresponding tar-archive.
}
\description{
This dataset contains 96 MALDI-TOF mass spectra of four different bacteria
species. Each species is represented by eight individual samples and each
sample has three technical replicates.
}
\examples{
library("MALDIquantExamples")
getPathSpecies()
}
\references{
This dataset was kindly provided by
Dr. Bryan R. Thoma \email{bryanthoma@yahoo.com}.
}
\keyword{datasets}

