% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{MALDIquantExamples-package}
\alias{MALDIquantExamples-package}
\title{Examples for MALDIquant}
\description{
This package provides example scripts and data to demonstrate the
usage of MALDIquant.
}
\details{
\tabular{ll}{
Package: \tab MALDIquantExamples \cr
License: \tab GPL (>= 3)\cr
URL: \tab http://www.github.com/sgibb/MALDIquantExamples/ \cr
}
}
\author{
Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
\url{http://www.github.com/sgibb/MALDIquantExamples/}
}
\keyword{package}

