% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotImsSlice}
\alias{plotImsSlice}
\title{Plot MALDI-IMS data}
\usage{
plotImsSlice(x, range, main)
}
\arguments{
\item{x}{list of \code{\linkS4class{MassSpectrum}} objects.}

\item{range}{\code{double}, length two, mass range}

\item{main}{\code{character} plot title.}
}
\value{
Nothing, plots an image.
}
\description{
This function is a basic function for MALDI-IMS data.
}
\seealso{
\url{https://github.com/sgibb/ims-shiny/}
}

